refactoring to support chat initialization
