+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/redux/action/ChatAction.kt
@@ -3,13 +3,12 @@
 
 package com.azure.android.communication.ui.chat.redux.action
 
-import com.azure.android.communication.ui.chat.error.ChatStateError
 import com.azure.android.communication.ui.chat.models.ParticipantInfoModel
 
 internal sealed class ChatAction : Action {
     class Initialization : ChatAction()
     class Initialized : ChatAction()
-    class Error(val chatStateError: ChatStateError) : ChatAction()
+    class Error() : ChatAction()
     class LocalParticipantInfo(val localParticipantInfoModel: ParticipantInfoModel) : ChatAction()
     class ChatThreadID(val chatThreadId: String) : ChatAction()
 }
