+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/redux/middleware/ChatActionHandler.kt
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.ui.chat.redux.middleware.listener
+package com.azure.android.communication.ui.chat.redux.middleware
 
 import com.azure.android.communication.ui.chat.error.ChatStateError
 import com.azure.android.communication.ui.chat.error.ErrorCode
@@ -11,13 +11,14 @@ import com.azure.android.communication.ui.chat.redux.action.ErrorAction
 import com.azure.android.communication.ui.chat.redux.state.ReduxState
 import com.azure.android.communication.ui.chat.service.ChatService
 
-internal class ChatActionListener(private val chatService: ChatService) {
+internal class ChatActionHandler(private val chatService: ChatService) {
     fun initialization(store: Store<ReduxState>) {
         try {
-            chatService.init()
+            chatService.initialize()
+            store.dispatch(ChatAction.Initialized())
         } catch (ex: Exception) {
             val error = ChatStateError(errorCode = ErrorCode.CHAT_JOIN_FAILED)
-            store.dispatch(ChatAction.Error(chatStateError = error))
+            store.dispatch(ChatAction.Error())
             store.dispatch(ErrorAction.ChatStateErrorOccurred(chatStateError = error))
         }
     }
@@ -26,7 +27,7 @@ internal class ChatActionListener(private val chatService: ChatService) {
         // TODO:subscribe to notifications
     }
 
-    fun chatError(store: Store<ReduxState>) {
+    fun onError(store: Store<ReduxState>) {
         // TODO:notify Contoso with Error
         // TODO:display UI discussion required
     }
