+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/redux/middleware/ChatServiceListener.kt
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.android.communication.ui.chat.redux.middleware.listener
+package com.azure.android.communication.ui.chat.redux.middleware
 
-import com.azure.android.communication.ui.chat.redux.middleware.ChatMiddleware
 import com.azure.android.communication.ui.chat.service.ChatService
 
 internal class ChatServiceListener(
