+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/service/ChatService.kt
@@ -7,7 +7,7 @@ import com.azure.android.communication.ui.chat.service.sdk.ChatSDK
 import com.azure.android.communication.ui.chat.service.sdk.wrapper.ChatMessageType
 
 internal class ChatService(private val chatSDK: ChatSDK) {
-    fun init() = chatSDK.init()
+    fun initialize() = chatSDK.init()
     fun sendMessage(
         type: ChatMessageType,
         content: String,
